+++ schema/accepted_schema.sql
@@ -2,9 +2,9 @@
 -- Do not edit this file directly. Instead, start a server and run "yarn makemigrations"
 -- as described in the README. This file should nevertheless be checked in to version control.
 --
--- Overall schema hash: 6d8270f44805a1ee0b363924964776e7
+-- Overall schema hash: 2e28d7576d143428c88b6c5a8ece4690
 
--- Accepted on 2023-11-21T12:12:51.000Z by 20231121T121251.add_givingSeason2023DonatedFlair.ts
+-- Accepted on 2023-11-24T20:09:25.000Z by 20231124T200925.move_on_connect_queries_into_migration.ts
 
 -- Extension "btree_gin", hash: 8833dd03dc128e186d2c560d248354fb
 CREATE EXTENSION IF NOT EXISTS "btree_gin" CASCADE;
@@ -1356,4 +1356,164 @@ CREATE TABLE "Votes" (
     "createdAt" timestamptz DEFAULT CURRENT_TIMESTAMP,
     "legacyData" jsonb
 );
+
+-- Function, hash: b46d62a3f1df590b2b32db6f51aa876d
+CREATE OR REPLACE FUNCTION fm_build_nested_jsonb(
+    target_path TEXT[],
+    terminal_element JSONB
+  )
+    RETURNS JSONB LANGUAGE sql IMMUTABLE AS
+   'SELECT JSONB_BUILD_OBJECT(
+      target_path[1],
+      CASE
+        WHEN CARDINALITY(target_path) = 1 THEN terminal_element
+        ELSE fm_build_nested_jsonb(
+          target_path[2:CARDINALITY(target_path)],
+          terminal_element
+        )
+      END
+    );'
+  ;
+
+-- Function, hash: 30fda1b23a6dad42fa6f09da3c46b6e9
+CREATE OR REPLACE FUNCTION fm_add_to_set(ANYARRAY, ANYELEMENT)
+    RETURNS ANYARRAY LANGUAGE sql IMMUTABLE AS
+   'SELECT CASE WHEN ARRAY_POSITION($1, $2) IS NULL THEN $1 || $2 ELSE $1 END;'
+  ;
+
+-- Function, hash: 73ea4de0fde6ca0eaf892ec947fab593
+CREATE OR REPLACE FUNCTION fm_add_to_set(
+    base_field JSONB,
+    target_path TEXT[],
+    value_to_add ANYELEMENT
+  )
+    RETURNS JSONB LANGUAGE sql IMMUTABLE AS
+   'SELECT CASE
+    WHEN base_field #> target_path IS NULL
+      THEN COALESCE(base_field, ''{}''::JSONB) || fm_build_nested_jsonb(
+        target_path,
+        JSONB_BUILD_ARRAY(value_to_add)
+      )
+    WHEN EXISTS (
+      SELECT *
+      FROM JSONB_ARRAY_ELEMENTS(base_field #> target_path) AS elem
+      WHERE elem = TO_JSONB(value_to_add)
+    )
+      THEN base_field
+    ELSE JSONB_INSERT(
+      base_field,
+      (SUBSTRING(target_path::TEXT FROM ''(.*)}.*$'') || '', -1}'')::TEXT[],
+      TO_JSONB(value_to_add),
+      TRUE
+    )
+    END;'
+  ;
+
+-- Function, hash: 47cabadc75b8331f567a50b081b94bae
+CREATE OR REPLACE FUNCTION fm_post_tag_similarity(
+    post_id_a TEXT,
+    post_id_b TEXT
+  )
+    RETURNS FLOAT LANGUAGE sql IMMUTABLE AS
+   'SELECT
+      COALESCE(SUM(LEAST(a, b))::FLOAT / SUM(GREATEST(a, b))::FLOAT, 0) AS similarity
+    FROM (
+      SELECT
+        GREATEST((a."tagRelevance"->"tagId")::INTEGER, 0) AS a,
+        GREATEST((b."tagRelevance"->"tagId")::INTEGER, 0) AS b
+      FROM (
+        SELECT JSONB_OBJECT_KEYS("tagRelevance") AS "tagId"
+        FROM "Posts"
+        WHERE "_id" IN (post_id_a, post_id_b)
+      ) "allTags"
+      JOIN "Posts" a ON a."_id" = post_id_a
+      JOIN "Posts" b ON b."_id" = post_id_b
+    ) "tagRelevance";'
+  ;
+
+-- Function, hash: 41a9827463ed86fb66c5877ce8fd2e6d
+CREATE OR REPLACE FUNCTION fm_jsonb_subset(target jsonb, candidate jsonb)
+  RETURNS BOOLEAN AS $$
+  DECLARE
+    key text;
+  BEGIN
+    FOR key IN SELECT jsonb_object_keys(candidate)
+    LOOP
+      IF NOT (target ? key AND target->>key = candidate->>key) THEN
+        RETURN FALSE;
+      END IF;
+    END LOOP;
+
+    RETURN TRUE;
+  END;
+  $$ LANGUAGE plpgsql;
+  ;
+
+-- Function, hash: 428d487b3e1991b66366492a805f7408
+CREATE OR REPLACE FUNCTION fm_post_tag_ids(post_id TEXT)
+    RETURNS TEXT[] LANGUAGE sql IMMUTABLE AS
+   'SELECT ARRAY_AGG(tags."tagId")
+    FROM "Posts" p
+    JOIN (
+      SELECT JSONB_OBJECT_KEYS("tagRelevance") AS "tagId"
+      FROM "Posts"
+      WHERE "_id" = post_id
+    ) tags ON p."_id" = post_id
+    WHERE (p."tagRelevance"->tags."tagId")::INTEGER >= 1;'
+  ;
+
+-- Function, hash: 005a9177ae69ef02fad0868493928c49
+CREATE OR REPLACE FUNCTION fm_confidence_sort(
+    ups INTEGER,
+    downs INTEGER,
+    downvote_multiplier FLOAT DEFAULT 1
+  ) RETURNS FLOAT LANGUAGE PLPGSQL IMMUTABLE AS $$
+  DECLARE
+    n INTEGER;
+    z FLOAT;
+    p FLOAT;
+    l FLOAT;
+    r float;
+    u FLOAT;
+  BEGIN
+    n := ups + (downs * downvote_multiplier);
+    IF n = 0 THEN
+      RETURN n;
+    END IF;
+    z := 1.281551565545;
+    p := ups::FLOAT / n::FLOAT;
+    l := p + 1 / (2 * n) * z * z;
+    r := z * SQRT(p * (1 - p) / n + z * z / (4 * n * n));
+    u := 1 + 1 / n * z * z;
+    RETURN (l - r) / u;
+  END $$
+  ;
+
+-- Function, hash: 1d0e667eb1ca49274453f3334fa5931f
+CREATE OR REPLACE FUNCTION fm_comment_confidence(
+    comment_id TEXT,
+    downvote_multiplier FLOAT DEFAULT 1
+  ) RETURNS FLOAT LANGUAGE sql AS $$
+    SELECT
+      fm_confidence_sort(
+        COALESCE(
+          SUM(v."power") FILTER (WHERE v."voteType" IN ('bigUpvote', 'smallUpvote')),
+          0
+        )::INTEGER,
+        COALESCE(
+          -SUM(v."power") FILTER (WHERE v."voteType" IN ('bigDownvote', 'smallDownvote')),
+          0
+        )::INTEGER,
+        downvote_multiplier
+      )
+    FROM "Comments" c
+    JOIN "Votes" v ON
+      v."documentId" = c."_id" AND
+      v."collectionName" = 'Comments' AND
+      v."isUnvote" IS NOT TRUE AND
+      v."cancelled" IS NOT TRUE AND
+      v."extendedVoteType" IS NULL
+    WHERE c."_id" = comment_id;
+  $$
+  ;
