+++ schema/schema_changelog.json
@@ -418,5 +418,10 @@
     "acceptsSchemaHash": "6d8270f44805a1ee0b363924964776e7",
     "acceptedByMigration": "20231121T121251.add_givingSeason2023DonatedFlair.ts",
     "timestamp": "2023-11-21T12:12:51.000Z"
+  },
+  {
+    "acceptsSchemaHash": "2e28d7576d143428c88b6c5a8ece4690",
+    "acceptedByMigration": "20231124T200925.move_on_connect_queries_into_migration.ts",
+    "timestamp": "2023-11-24T20:09:25.000Z"
   }
 ]
